// packages/aws-cli/src/AwsCli/types.ts
export interface AwsCliBaseProps {
  service_name: string
  service_description: string
  service_url: string
  service_commands: [
    {
      command_name: string
      command_url: string
      command_description?: string
      command_synopsis?: string
      command_options?: string[]
      command_options_description?: string
      command_output?: string
      command_examples?: string
    },
  ]
}

export interface iSelectionData {
  service_name: string
  command_name: string
  command_options: {
    command_options_name: string
    command_options_value: string
  }[]
}

export interface iHandleValueChange {
  event: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
  command_option: string
}